Converter: Organize controls and logic
