+++ src/components/Migrate/Converter/ConverterForm.tsx
@@ -1,28 +1,20 @@
-import React, { useCallback, useMemo, useState } from 'react'
+import React, { useMemo } from 'react'
 import { css } from 'styled-components'
 import {
-  TextInput,
-  Link,
   useTheme,
   useLayout,
-  Info,
   GU,
-  noop,
   // @ts-ignore
 } from '@aragon/ui'
 // @ts-ignore
 import TokenAmount from 'token-amount'
-import BrandButton from '../../BrandButton/BrandButton'
 import { fontWeight } from '../../../style/font'
-import { TokenConversionType, ValidationStatus } from '../types'
+import { TokenConversionType } from '../types'
 import { useMigrateState } from '../MigrateStateProvider'
 import { shadowDepth } from '../../../style/shadow'
 import { useAccountBalances } from '../../../providers/AccountBalances'
-import { parseUnits } from '../../../utils/math-utils'
-import { BigNumber } from 'ethers'
-import ConverterButton from './ConverterButton'
+import ConverterFormControls from './ConverterFormControls'
 
-const BLOG_POST_URL = ''
 const AMOUNT_DIGITS = 6
 const TOKEN_SYMBOL: Record<TokenConversionType, string> = {
   ANT: 'ANT',
@@ -119,142 +111,13 @@ function ConverterForm(): JSX.Element {
           grid-area: inputs;
         `}
       >
-        <FormControls />
+        <ConverterFormControls
+          tokenSymbol={tokenSymbol}
+          amountDigits={AMOUNT_DIGITS}
+        />
       </div>
     </form>
   )
-}
-
-function FormControls() {
-  const [amount, setAmount] = useState('')
-  const theme = useTheme()
-  const { continueToSigning } = useMigrateState()
-  const { layoutName } = useLayout()
-  const { conversionType } = useMigrateState()
-  const { formattedAmount, validationStatus } = useInputValidation(
-    amount,
-    AMOUNT_DIGITS
-  )
-
-  console.log(formattedAmount, validationStatus)
-
-  const stackedButtons = layoutName === 'small'
-  const tokenSymbol = TOKEN_SYMBOL[conversionType]
-
-  const handleAmountChange = useCallback((event) => {
-    setAmount(event.target.value)
-  }, [])
-
-  return (
-    <>
-      <label
-        css={`
-          display: block;
-        `}
-      >
-        <h3
-          css={`
-            font-weight: ${fontWeight.medium};
-            margin-bottom: ${1 * GU}px;
-          `}
-        >
-          Enter the amount you would like to convert
-        </h3>
-        <TextInput
-          wide
-          placeholder={`0.0 ${tokenSymbol} v1`}
-          value={amount}
-          onChange={handleAmountChange}
-          type="number"
-          css={`
-            display: block;
-          `}
-        />
-      </label>
-      <p
-        css={`
-          margin-top: ${1 * GU}px;
-          color: ${theme.surfaceContentSecondary};
-        `}
-      >
-        You will receive:{' '}
-        <span css={`font-weight: ${fontWeight.medium}; color ${theme.accent}`}>
-          {formattedAmount}
-        </span>{' '}
-        {tokenSymbol} v2
-      </p>
-      <Info
-        css={`
-          margin-top: ${3 * GU}px;
-          margin-bottom: ${2 * GU}px;
-        `}
-      >
-        Please read our{' '}
-        <Link href={BLOG_POST_URL}>ANT Migration blog post</Link> if you have
-        any questions.
-      </Info>
-      <div
-        css={`
-          display: grid;
-          grid-gap: ${1 * GU}px;
-
-          grid-template-columns: ${stackedButtons ? 'auto' : '1fr 1fr'};
-        `}
-      >
-        <BrandButton wide>Back</BrandButton>
-        <ConverterButton
-          // TODO: Add call to account module
-          onConnect={noop}
-          onContinue={continueToSigning}
-          status={validationStatus}
-        />
-      </div>
-    </>
-  )
-}
-
-function useInputValidation(amount: string, digits: number) {
-  const { antV1 } = useAccountBalances()
-  const { balance, decimals } = antV1
-
-  const parsedAmountBn = useMemo(() => parseInputValue(amount, decimals), [
-    amount,
-    decimals,
-  ])
-
-  const formattedAmount = useMemo((): string => {
-    return new TokenAmount(parsedAmountBn, decimals).format({
-      digits,
-    })
-  }, [parsedAmountBn, decimals, digits])
-
-  const validationStatus = useMemo((): ValidationStatus => {
-    if (!balance) {
-      return 'notConnected'
-    }
-
-    if (parsedAmountBn.gt(balance)) {
-      return 'insufficientBalance'
-    }
-
-    if (parsedAmountBn.isZero()) {
-      return 'noAmount'
-    }
-
-    return 'valid'
-  }, [parsedAmountBn, balance])
-
-  return {
-    parsedAmountBn,
-    formattedAmount,
-    validationStatus,
-  }
-}
-
-function parseInputValue(inputValue: string, decimals: number): BigNumber {
-  const trimmedValue = inputValue.trim()
-
-  return parseUnits(trimmedValue || '0', decimals)
 }
 
 export default ConverterForm
